+++ demos/chinook/src/test/java/is/codion/framework/demos/chinook/domain/ChinookTest.java
@@ -18,6 +18,7 @@
  */
 package is.codion.framework.demos.chinook.domain;
 
+import is.codion.common.db.exception.DatabaseException;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.demos.chinook.domain.Chinook.Playlist.RandomPlaylistParameters;
 import is.codion.framework.demos.chinook.domain.impl.ChinookImpl;
@@ -37,10 +38,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ChinookTest extends DomainTest {
 
-	private static final ChinookImpl DOMAIN = new ChinookImpl();
-
 	public ChinookTest() {
-		super(DOMAIN, new ChinookEntityFactory());
+		super(new ChinookImpl(), ChinookEntityFactory::new);
 	}
 
 	@Test
@@ -122,12 +121,12 @@ public class ChinookTest extends DomainTest {
 
 	private static final class ChinookEntityFactory extends DefaultEntityFactory {
 
-		private ChinookEntityFactory() {
-			super(DOMAIN.entities());
+		private ChinookEntityFactory(EntityConnection connection) {
+			super(connection);
 		}
 
 		@Override
-		public void modify(Entity entity) {
+		public void modify(Entity entity) throws DatabaseException {
 			super.modify(entity);
 			if (entity.entityType().equals(Album.TYPE)) {
 				entity.put(Album.TAGS, asList("tag_one", "tag_two", "tag_three"));
@@ -135,7 +134,7 @@ public class ChinookTest extends DomainTest {
 		}
 
 		@Override
-		protected <T> T value(Attribute<T> attribute) {
+		protected <T> T value(Attribute<T> attribute) throws DatabaseException {
 			if (attribute.equals(Album.TAGS)) {
 				return (T) asList("tag_one", "tag_two");
 			}