+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertEmployee.java
@@ -33,7 +33,7 @@ public final class InsertEmployee extends AbstractPerformer {
 		SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
 		selectRandomRow(departmentModel.tableModel());
 		SwingEntityModel employeeModel = departmentModel.detailModel(Employee.TYPE);
-		Entity employee = new DefaultEntityFactory(application.entities()).entity(Employee.TYPE);
+		Entity employee = new DefaultEntityFactory(application.connection()).entity(Employee.TYPE);
 		employee.put(Employee.DEPARTMENT_FK, departmentModel.tableModel().selectionModel().getSelectedItem());
 		employeeModel.editModel().set(employee);
 		employeeModel.editModel().insert();